+++ WarOfMini/Client/Codes/PlayerCamera.cpp
@@ -94,9 +94,10 @@ void CPlayerCamera::FixMouse(void)
 	POINT ptMouse;
 	ZeroMemory(&ptMouse, sizeof(POINT));
 
-	ptMouse = { WINCX >> 1, WINCY >> 1 };
+	ptMouse = { (long)WINCX >> 1, (long)WINCY >> 1 };
 
 	ClientToScreen(g_hWnd, &ptMouse);
+
 	SetCursorPos(ptMouse.x, ptMouse.y);
 }
